feat: temporary directory guesser
Some PHP runtime environments (e.g. shared hosters) are (mis-)configured
in a way that the ini-directive sys_temp_dir points to a folder that is
not actually accessible. For those hosts `TemporaryFileWriterGuesser`
tries a fewpossible sources to find a proper temporary directory. To make
this behavior the default, `Twint\Sdk\Client` constructor now takes either
a factory creating a `FileWriter` (`callable(): FileWriter`) or a
`FileWriter` instance directly.